Address feedback + authenticator interface
