+++ src/Microsoft.SqlTools.Authentication/Microsoft.SqlTools.Authentication.csproj
@@ -9,7 +9,6 @@
 		<PackageReference Include="Microsoft.Data.SqlClient" />
 		<PackageReference Include="Microsoft.Identity.Client" />
 		<PackageReference Include="Microsoft.Identity.Client.Extensions.Msal" />
-		<PackageReference Include="Newtonsoft.Json" />
 		<PackageReference Include="System.Configuration.ConfigurationManager" />
 	</ItemGroup>
