+++ src/Microsoft.SqlTools.Authentication/Sql/AuthenticationProvider.cs
@@ -4,8 +4,6 @@
 //
 
 using Microsoft.Data.SqlClient;
-using Microsoft.SqlTools.Authentication.Utility;
-using Microsoft.SqlTools.Utility;
 
 namespace Microsoft.SqlTools.Authentication.Sql
 {
@@ -17,11 +15,9 @@ namespace Microsoft.SqlTools.Authentication.Sql
     /// </summary>
     public class AuthenticationProvider : SqlAuthenticationProvider
     {
-        private const string ApplicationClientId = "a69788c6-1d43-44ed-9ca3-b83e194da255";
-        private const string AzureTokenFolder = "Azure Accounts";
         private const string s_defaultScopeSuffix = "/.default";
 
-        private Authenticator authenticator;
+        private IAuthenticator authenticator;
 
         /// <summary>
         /// Instantiates AuthenticationProvider to be used for AAD authentication with MSAL.NET and MSAL.js co-ordinated.
@@ -29,22 +25,9 @@ namespace Microsoft.SqlTools.Authentication.Sql
         /// <param name="applicationName">Application Name that identifies user folder path location for reading/writing to shared cache.</param>
         /// <param name="applicationPath">Application Path directory where application cache folder is present.</param>
         /// <param name="authCallback">Callback that handles AAD authentication when user interaction is needed.</param>
-        public AuthenticationProvider(string applicationName, string applicationPath, string cacheName, MSALEncryptedCacheHelper.IVKeyReadCallback callback)
+        public AuthenticationProvider(IAuthenticator authenticator)
         {
-            if (string.IsNullOrEmpty(applicationName))
-            {
-                applicationName = nameof(SqlTools);
-                Logger.Warning($"Application Name not received with command options, using default application name as: {applicationName}");
-            }
-
-            if (string.IsNullOrEmpty(applicationPath))
-            {
-                applicationPath = Utils.BuildAppDirectoryPath();
-                Logger.Warning($"Application Path not received with command options, using default application path as: {applicationPath}");
-            }
-
-            var cachePath = Path.Combine(applicationPath, applicationName, AzureTokenFolder);
-            this.authenticator = new Authenticator(new (ApplicationClientId, applicationName, cachePath, cacheName), callback);
+            this.authenticator = authenticator;
         }
 
         /// <summary>
