+++ src/Microsoft.SqlTools.Shared/Utility/Constants.cs
@@ -15,6 +15,9 @@ namespace Microsoft.SqlTools.Shared.Utility
         public const string ActiveDirectoryInteractive = "ActiveDirectoryInteractive";
         public const string ActiveDirectoryPassword = "ActiveDirectoryPassword";
 
+        // Azure authentication (MSAL) constants
+        public const string ApplicationClientId = "a69788c6-1d43-44ed-9ca3-b83e194da255";
+        public const string AzureTokenFolder = "Azure Accounts";
         public const string AzureAccountProviderCredentials = "azureAccountProviderCredentials";
         public const string MsalCacheName = "accessTokenCache";
     }
